# docs/swagger.yaml
host: localhost:8080
info:
  contact:
    name: API Support
  description: This is a sample server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /{modelName}/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: Model Name
        in: path
        name: modelName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: CRUD FindMany
...
  /users/token:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: User GetToken
swagger: "2.0"